Add HASHES, MERGES to equality operator
There is no good reason to disallow hash and merge on = it was just
not considered on the initial implementation.

The version migration however is kinda tricky, as postgres doesn't
allow to `ALTER OPERATOR` according to [hackers](https://github.com/postgres/postgres/commit/321eed5f0f7563a0cabb3d7a98132856287c1ad1) because of query plan caching.

```
Other options cannot be changed, as it's not totally clear if cached plans
would need to be invalidated if one of the other options change.
```

An upgrade is probably not risky as it just adds more execution options to
the planer.
A downgrade however could result in executing a hash/merge join that actually isn't
allowed.
This might not be a problem as well as the previous version has all the needed
functionality to execute a hash/merge join (i.e. the underlying procedure didn't change).